Allow sharing configuration between all spaces
This change allows operators to create a configuration file that will be applied to all spaces in the Cockpit instance by creating the file `config/shared.php` with the same format as the `config/config.php`. If the file exists, it will be loaded as a base and the space specific configuration is applied on top of it.